requirements.txt
click==8.1.7
importlib-metadata==6.7.0
kafka-python==2.0.2
typing_extensions==4.7.1
zipp==3.15.0
